Implement raw query execution
