+++ src/Database/Adapter/SQL.php
@@ -36,8 +36,25 @@ abstract class SQL extends Adapter
      */
     public function ping(): bool
     {
-        return $this->getPDO()
+        $query = $this->getPDO()
             ->prepare("SELECT 1;")
+            ->queryString;
+
+        return $this->execute($query);
+    }
+
+    /**
+     * Execute raw command
+     * @param mixed $query
+     *
+     * @return mixed
+     * @throws Exception
+     * @throws PDOException
+     */
+    public function execute(mixed $query): mixed
+    {
+        return $this->getPDO()
+            ->prepare($query)
             ->execute();
     }
